optimized some code parts, changed starting path from "/index.html" to "/" only moved noAvatar.png to "uploads" folder
users can now click on the saved products in their profile and are redirected to the product page

improved photo upload function, every user can now permanently upload 1 photo at a time
-the photo will be deleted if it is changed to a default photo or gets removes via remove option

therefore, added functions in server.js

and some more..

+++ web/server.js
@@ -8,11 +8,11 @@ const multer = require("multer");
 
 const storage = multer.diskStorage({
     destination: function (req, file, cb) {
-        clearDirectory();
+        clearDirectory(req.session.user);
         cb(null, 'files/resources/uploads/')
     },
     filename: function (req, file, cb) {
-        cb(null, Date.now() + path.extname(file.originalname)) //Appending extension
+        cb(null, "user" + req.session.user + " - " + Date.now() + path.extname(file.originalname)) //Appending extension
     }
 })
 const upload = multer(
@@ -61,17 +61,19 @@ app.use('/api', wheelchairRouter);
 app.use('/api', accountRouter);
 
 app.post("/uploadFiles", upload.single("file"), sendFileUrl);
+app.delete("/deleteUserFile", deleteUserUpload);
+app.get("/readLibrary", sendImgLibrary);
 
 app.listen(port, (error) => {
     if (error) {
         console.log(error);
     } else {
         //added "/index.html" at the end so that document location pathname could be used to determine what the current page is
-        console.log(`Server listening at http://localhost:${port}/index.html`)
+        console.log(`Server listening at http://localhost:${port}`)
     }
 });
 
-function clearDirectory() {
+function clearDirectory(user) {
 
     const directory = "files/resources/uploads";
     fs.readdir(directory, (err, files) => {
@@ -79,14 +81,67 @@ function clearDirectory() {
 
         for (const file of files) {
 
-            fs.unlink(path.join(directory, file), err => {
-                if (err) throw err;
-            });
+            if (file.includes(user)) {
+                fs.unlink(path.join(directory, file), err => {
+                    if (err) throw err;
+                });
+            }
         }
     });
 }
+
 function sendFileUrl(req, res) {
 
     let url = req.file.path.split("resources").pop();
+
     res.status(200).send({url: url});
+}
+
+function deleteUserUpload(req, res) {
+
+    try {
+        clearDirectory(req.session.user)
+
+        setTimeout(() => res.send({bo: true}), 20);
+
+    } catch (e) {
+
+        res.status(500).send({bo: null});
+    }
+}
+
+function sendImgLibrary(req, res) {
+
+    const rootDirectory = "files/resources";
+    let array = [];
+
+    fs.readdir(`${rootDirectory}/uploads`, (err, files) => {
+        if (err) throw err;
+
+        for (const file of files) {
+
+            let filePath = `/uploads/:${file}`
+            array.push(filePath)
+        }
+
+    });
+
+    fs.readdir(`${rootDirectory}/wheelchairs`, (err, subDir) => {
+        if (err) throw err;
+
+        for (const dir of subDir) {
+
+            fs.readdir(`${rootDirectory}/wheelchairs/${dir}`, (err, files) => {
+                if (err) throw err;
+
+                for (const file of files) {
+
+                    let filePath = `/wheelchairs/${dir}/:${file}`
+                    array.push(filePath)
+                }
+
+            });
+        }
+    });
+    setTimeout(() => res.send({array: array}), 20);
 }
